towards hlint and stack build --pedantic
removed some hlint issues
removed some stack build --pedantic issues in defs

Start Development Acceptance Test
End Development Acceptance Test
Start Development Acceptance Check
End Development Acceptance Check
Checking development acceptance test output for file 20171306_164135.11_DatOutpu
t.datr
- TestCopyright:         PASS
- Test build:            PASS
- Torxakis build:        PASS
- SelfTests:             PASS
- TestExamps:
--- Stimulus Response Tests
----- StimulusResponse Test 1: PASS
----- StimulusResponse Test 2: PASS
----- StimulusResponse Test 3: PASS
----- StimulusResponse Test 4: PASS
----- StimulusResponse Test 5: PASS
--- CustomerOrders Tests
----- CustomersOrders Test: PASS
--- LuckyPeople Tests:
----- LuckyPeopleTest Test 1: PASS
----- LuckyPeopleTest Test 2: PASS
Checking finished

+++ sys/defs/src/Params.hs
@@ -71,12 +71,12 @@ initParams  =  Map.fromList $ map ( \(x,y,z) -> (x,(y,z)) )
 -- test/simulation input generation
   
   , ( "param_Test_inputEager"    , "3"         , \s ->   all Char.isDigit s
-                                                      && 0 <= read s && read s <= 3 )
+                                                      && 0 <= (read s::Integer) && (read s::Integer) <= 3 )
             -- param_Test_inputEager :: Int (0 <= param_Test_inputEager <= 3)
             -- level of input eagerness: 0 .. 3
 
   , ( "param_Sim_inputEager"     , "0"         , \s ->   all Char.isDigit s 
-                                                      && 0 <= read s && read s <= 3 )
+                                                      && 0 <= (read s::Integer) && (read s::Integer) <= 3 )
             -- param_Sim_inputEager :: Int (0 <= param_Test_inputEager <= 3)
             -- level of input eagerness: 0 .. 3
